+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Action for compare a script with an environment (context menu or Alt+C)
+- Auto-detection of script interfaces
 
 ## [1.2.5] - 2023-02-03
 - Support for IntelliJ 2022.3